Added push notification object and catching to the app. Fixed some warnings.

+++ Huntr/Classes/Controllers/TeamsListViewController.m
@@ -483,7 +483,7 @@
         GameViewController * gameViewController = segue.destinationViewController;
         gameViewController.selectedGame = self.selectedGame;
         
-        NSLog(@"%u",self.selectedGame.status);
+        NSLog(@"%ld",(long)self.selectedGame.status);
     }
 }
